+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-16 22:05:43
-// Commit: test(theme): implement widget integration tests for theme
+// Updated at: 2026-08-18 13:33:07
+// Commit: test(theme): add unit tests for theme business logic
 
 class themeComponentState {
-  final String commitRef = '3660';
-  final String lastMessage = 'test(theme): implement widget integration tests for theme';
-  final DateTime timestamp = DateTime.parse('2026-08-16T22:05:43+05:00');
+  final String commitRef = '3693';
+  final String lastMessage = 'test(theme): add unit tests for theme business logic';
+  final DateTime timestamp = DateTime.parse('2026-08-18T13:33:07+05:00');
 }