+++ CHANGELOG.md
@@ -3694,3 +3694,4 @@
 - [2026-08-18 12:13] test(ui): add unit tests for ui business logic
 - [2026-08-18 12:50] feat(network): implement shimmer loading skeleton in network
 - [2026-08-18 13:09] test(core): implement widget integration tests for core
+- [2026-08-18 13:33] test(theme): add unit tests for theme business logic